pyproject.toml
[project]
name = "fastapi-scope-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pydantic>=2.10.4",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.36",
]
